Enable gs mask checking
This commit will modify movq_rm to check whether the destination memory
address has a base address that is equivalent to the base address stored
in the gs register. This is a debug change to ensure the region specific
zone allocations are working properly. Currently, this commit causes
Firefox to crash since movq_rm tries to write to the stack (which is
outside the zone sandbox).