Contiguously allocate memory in sandbox
This commit will update all JavaScript memory allocations (Nursery,
Tenured Heap) to be within the sandbox. Currently, the sandbox is the
4GB region of memory starting at `0x0000000100000000`. Note that there
are still some JavaScript objects allocated through the BaseAllocator
that will have to be manually moved into the sandbox on a case by case
basis. Furthermore, there will be issues with ArrayBuffers since they
will interfere with the sandbox bump allocator (which expects
allocations to be multiples of ChunkSize).