Move Nursery position pointer into the sandbox.
This commit moves the position_ pointer which is a bump pointer used by
the Nursery into the sandbox. This is because the JIT can allocate
itself memory and needs to be able to modify the position_ pointer. With
this change, the blank hello_test.html page loads with no (heap)
violations.