Chrome UI sandboxed
This commit allows a website that runs no Javascript other than the
Chrome UI to run successfully. It implements an interface for the JIT
compiler to use a trusted C++ function to modify the realm_ pointer. This
trusted C++ function validates that the realm_ pointer always points to
a valid realm.