Chunk based allocator for JitSandbox heap
Adding an allocator for the 4GB JitSandbox Heap region. The allocator
tracks free/allocated GC Chunks which it treats as the unit of
allocation.

This works currently assumes the jemalloc chunks size is
also the same as the JS GC chunk size since the same underlying
allocator is used for the Sandbox Malloc arena managed by jemalloc.

With this change the GC can properly release pages back to the
JitSandbox heap. Sandbox Malloc arena frees are still not properly
handled and will leak chunks.