Use MC::NullHandleValue instead of JS::NullHandleValue
In Monkeycage configurations where SpiderMonkey is separately compiled
and loaded by a sandboxing runtime - global constants such as JS::NullHandleValue
are not available/resolvable at link-time - since the sandbox is may be created
at a dynamic address.

Handling only the case of JS::NullHandleValue by exposing the address of the
global constant JS::JSVAL_NULL as a new JS_SANDBOX only JSAPI and re-creating
the HandleValue in the shim library as MC::NullHandleValue.

When Monkeycage is disabled / used in the no-op configuration MC::NullHandleValue
just returns JS::NullHandleValue.