Expose SpiderMonkey global data through Monkeycage API
When Monkeycage is enabled - especially in one of the dynamic
configurations - external code cannot directly link to on global/static
constant symbols or their addresses. These symbols need to be exposed
through new explicit getter functions added to JSAPI.

- In Gecko code we use Monkeycage wrappers around these constants so that
the code can be oblivious to whether the sandbox is actually enabled or not.
When disabled, these wrapper getters defer to the original symbols. Otherwise
they retrieve them from the getter function and cache them to prevent
unnecessary transitions.

- The JS_SANDBOX_API define is added for code in Spidermonkey headers
when included by non Spidermonkey code when the sandbox is enabled.
This is used for instance when the above new getter functions need to be
used within the Spidermonkey headers under js/public/. We don't want to
use the new getters for code in Spidermonkey that includes the headers.

Ideally we should not overuse this and eventually create Monkeycage API
wrappers on any code that uses the getters.