Use Monkeycage BaseProxyHandler class and APIs in Gecko
We want to expose the entire BaseProxyHandler and Wrapper class hierarchy
through a bunch of intermediate glue classes because:

- Virtual Table symbols: We want all classes/instances of classes in the
Proxy Handler hierarchy that Firefox (SpiderMonekey or Gecko) will create
to use a common set of vtable symbols. This just makes things less messy
and allows the assumptions we can make about whether a virtual call will
require a sandbox transition or not to be directly determinable from the
type.

- Virtual Calls: Related to the above point, we need a point in the code
to "intercept" virtual function calls from within SpiderMonkey on
instances of ProxyHandler classes defined in Gecko. Since the virtual table
for these classes is populated automatically by the C++ compilation of Gecko
outside the reaches of LFI/the sandboxing toolchain, there is no automated way
of indicating this. The js::sandbox::* ProxyHandler classes make this explicit
by accepting a "vtable" of function pointers during construction where these
can be registered as sandbox callbacks using the usual mechanisms.

- Tainting: Re-defining the ProxyHandler class hierarchy in mc::* allows us
to potentially change the signatures of the virtual functions to be implemented
by inheriting classes in Gecko to have arguments with the Monkeycage Tainted types.