Use Monkeycage Wrapper class hierarchy and API
Switch all the Wrapper classes to use external glue classes that properly
handle virtual calls that need to transition from within SpiderMonkey to
the external implementation.

This does NOT currently handle two very critical bits:

- Tainting: The instance of the external Handler/Wrapper class hierarchy
is stored as an "app pointer" in the SpiderMonkey internal glue class. This
is not currently validated when retrieved.

- Gecko to SpiderMonkey Transitions: While the present glue works and allows
for all external wrappers to be retrieved and be treated uniformly, it does
not properly deal with the case when the actual Handler/Wrapper is one which has
had it's vtable populated within SpiderMonkey.